+++ lib/storage/src/tokens/storage-ttl-delay.ts
@@ -1,2 +1,2 @@
-// eslint-disable-next-line no-magic-numbers
+// eslint-disable-next-line @typescript-eslint/no-magic-numbers
 export const STORAGE_TTL_DELAY: number = 1000 * 60; // 1min